wince: trace cold path into kernel init and suspend
Tried: expanded the one-shot WinCE boot-path probe coverage from the existing pre-init milestones up through the cold kernel and scheduler pages, switched the path-probe mask to 64 bits, and logged the low-power branch inputs that feed the suspend helper.\n\nOutcome: a fresh NAND cold boot still times out in the suspend loop at 0x80079990, but the trace now shows the real cold path state at that point (PA2400=0x03020100, PA2408=0x03020100, PA251C=0x0000000F, PA2554=0x00000002, and the low-power inputs 0xAF000046, 0xAA0008A0, 0xAA000A00, 0xAA00130C all zero). NK_XLAT also proves execution reaches the cold kernel pages at 0x8007B57C, 0x800947C8, 0x8008B21C, 0x8007A3FC, and 0x80079898 before stalling. Linux regression still booted to userspace under timeout with Initializing RT netlink socket and Starting kswapd, and the WinCE-only probes stayed silent there.\n\nLearned: the current failure is no longer an early boot-dispatch or random-RAM problem. The emulator is taking a true cold path into kernel init and then falling into the scheduler/low-power flow, where interrupts become pending but do not produce forward progress out of the suspend helper.\n\nNext: trace the wake side of 0x80079910/0x80079990 and the surrounding VR41xx wait/interrupt delivery path to find why pending Cause.IP bits never advance execution past the suspend loop.

+++ src/wince_boot_types.h
@@ -34,7 +34,7 @@ typedef struct {
     /* Cold boot state */
     bool cold_boot_copy_done;
     uint32_t cold_boot_pc_probes_logged;
-    uint32_t boot_path_probe_mask;
+    uint64_t boot_path_probe_mask;
 
     /* Vector tracking */
     wince_vector_owner_t vector_owner;
